feat: specially handle `pkgs` input
If nixpkgs is declared in a flake as the `pkgs` input, then the
package collection will automatically be imported under the given
input for quick access, e.g. `pkgs.zlib` instead of
`pkgs.legacyPackages.zlib`.

In addition, the `lib` attribute will be brought into the top-level of
the functor arguments for quick reference, e.g.
`nosys inputs ({lib, pkgs, ...}: {})`